Change distance type of asset data model from float to integer

+++ app/GenAIFinder/app/src/main/java/io/krakau/genaifinder/service/api/model/data/Asset.kt
@@ -6,7 +6,7 @@ data class Asset(
     //@SerializedName("_id") val id: MongoId,
     val metadata: Metadata,
     val nnsId: Long,
-    val distance: Float
+    val distance: Int
     //@SerializedName("_class") val className: String
 )
